2016 - day 9

+++ 2016/day09/test.h
@@ -4,5 +4,5 @@
 
 namespace aoc2016_day09 {
     int part_1(std::string_view path);
-    int part_2(std::string_view path);
+    long long part_2(std::string_view path);
 }
